requirements.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Data Analysis
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
statsmodels==0.14.1

# Frontend
streamlit==1.30.0
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.17.0
altair==5.2.0

# Reports
reportlab==4.0.8
openpyxl==3.1.2
xlsxwriter==3.1.9
xlrd==2.0.1

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# Data Import
requests==2.31.0
eurostat==1.0.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
